+++ packages/react-geo/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @aws-amplify/ui-react-geo
 
+## 2.0.23
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @aws-amplify/ui-react-core@3.0.22
+
+## 2.0.22
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @aws-amplify/ui-react-core@3.0.21
+
 ## 2.0.21
 
 ### Patch Changes